requirements.txt
asgiref==3.6.0
cloudinary==1.32.0
Django==3.2.18
django-cloudinary-storage==0.3.0
django-countries==7.2.1
django-multiselectfield==0.1.12
Pillow==9.4.0
pytz==2022.7.1
sqlparse==0.4.3
